Split keymaster content routes
