+++ services/keymaster/server/src/keymaster-api.ts
@@ -16,16 +16,22 @@ import { InvalidParameterError } from '@didcid/common/errors';
 import config from './config.js';
 import { createAddressRouter } from './keymaster-address-router.js';
 import { createAddressTailRouter } from './keymaster-address-tail-router.js';
+import { createAssetRouter } from './keymaster-asset-router.js';
 import { createRequireAdminKey } from './keymaster-admin.js';
-import { createContentRouter } from './keymaster-content-router.js';
 import { createCoreRouter } from './keymaster-core-router.js';
 import { createCredentialRouter } from './keymaster-credential-router.js';
 import { createDidCommRouter } from './keymaster-didcomm-router.js';
+import { createDmailRouter } from './keymaster-dmail-router.js';
+import { createFileRouter } from './keymaster-file-router.js';
 import { createIdentityRouter } from './keymaster-identity-router.js';
+import { createImageRouter } from './keymaster-image-router.js';
 import { createLightningRouter } from './keymaster-lightning-router.js';
 import { createNostrRouter } from './keymaster-nostr-router.js';
+import { createNoticeRouter } from './keymaster-notice-router.js';
+import { createPollRouter } from './keymaster-poll-router.js';
 import { createPublicRouter } from './keymaster-public-router.js';
 import type { CreateKeymasterRouterOptions } from './keymaster-router-types.js';
+import { createVaultRouter } from './keymaster-vault-router.js';
 import promClient from 'prom-client';
 import pino from 'pino';
 import { pinoHttp } from 'pino-http';
@@ -171,7 +177,13 @@ v1router.use(createAddressTailRouter(routerOptions));
 v1router.use(createNostrRouter(routerOptions));
 v1router.use(createLightningRouter(routerOptions));
 v1router.use(createCredentialRouter(routerOptions));
-v1router.use(createContentRouter(routerOptions));
+v1router.use(createAssetRouter(routerOptions));
+v1router.use(createPollRouter(routerOptions));
+v1router.use(createImageRouter(routerOptions));
+v1router.use(createFileRouter(routerOptions));
+v1router.use(createVaultRouter(routerOptions));
+v1router.use(createDmailRouter(routerOptions));
+v1router.use(createNoticeRouter(routerOptions));
 
 app.use('/api/v1', v1router);
